Debug ftw segfault (#20)
Using reflect.value on the pointer to a handle type reference is not
correct since it will return the memory address of the temporary (e.g.
the argument). Similarly, taking the address of the transaction
interface type is also incorrect.

To make it easier to reason about, wrap transactions in a handle as well
and use unsafe.Pointer to the reference types.

---------

Signed-off-by: William Zhang <wtzhang23@gmail.com>